refactored my books page to organize components better
